Add scripts

+++ scripts/run_evaluation.sh
@@ -3,8 +3,14 @@
 chmod +x basex_commit.sh
 chmod +x evaluation_get.sh
 chmod +x evaluation_result.sh
-server_PID=$(<server_PID.txt)
-sleep 2
+evaluation_version=$@
+if [ $evaluation_version = "1" ]; then
+    ./basex_version_setup.sh /basex d1bb20b
+    sleep 2
+    server_PID=`cat ../../server_PID.txt`
+fi 
 mvn test -Dtest=EvaluationTest#evaluationTest$@ test
-kill -9 $server_PID
-sleep 2
+if [ $evaluation_version = "1" ]; then
+    kill -9 $server_PID
+    sleep 2
+fi
